requirements.txt
openai>=1.106.0
requests>=2.32.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
streamlit>=1.28.0
pandas>=2.3.0
plotly>=5.15.0
